+++ src/main/java/com/drpweb/disease/DiseaseService.java
@@ -4,7 +4,7 @@ package com.drpweb.disease;
  * Created by ADMIN on 8/27/2016.
  */
 public interface DiseaseService {
-     Disease getDiseaseById(int id);
+     Disease getDiseaseById(Long id);
      Disease findByDiseaseName(String diseaseName);
 
 }
